% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualize.R
\name{annualize}
\alias{annualize}
\alias{annualize.volatility}
\alias{annualize.variance}
\title{Annualize Volatility or Variance}
\usage{
annualize(x, ...)

\method{annualize}{volatility}(x, dib = 252, ...)

\method{annualize}{variance}(x, dib = 252, ...)
}
\arguments{
\item{x}{an object whose class will determine the method to be dispatched.
Volatility or variance, will become annual.}

\item{...}{further arguments to be passed to the next method.}

\item{dib}{days in one year, day count convention.}
}
\value{
Same class with annual volatility or variance.
}
\description{
Transforms variance or volatility to annual. There is no default
method, so only classes \code{"volatility"}, \code{"variance"}, are supported.
}
\examples{
annualize(as.variance(.5))
annualize(as.volatility(.25))

}
\seealso{
\code{\link{daily}} to transform variance or volatility to daily basis.
}
