% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{as.variance}
\alias{as.variance}
\alias{as.variance.default}
\alias{as.variance.volatility}
\title{variance Class Coercing}
\usage{
as.variance(x, ...)

\method{as.variance}{default}(x, annualized = FALSE, dib = NULL, ...)

\method{as.variance}{volatility}(x, ...)
}
\arguments{
\item{x}{an object whose class will determine the method to be dispatched.
Value is the volatility or variance.}

\item{...}{further arguments to be passed to the next method, such as:}

\item{annualized}{for default method, TRUE if variance value in x is annualized,
FALSE if variance is daily.}

\item{dib}{days in one year, day count convention.}
}
\value{
All values of x as \code{"variance"} class, according to given daily or
annualized type.
}
\description{
Generic function for coercing to \code{"variance"}. There is a default
method and a method for \code{"volatility"}.
}
\examples{
as.variance(1)
as.variance(as.volatility(.5))
as.variance(0.2, TRUE, 252)
as.variance(as.volatility(100:300 / 100, TRUE))

}
\seealso{
\code{\link{as.volatility}}.
}
