% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{as.volatility}
\alias{as.volatility}
\alias{as.volatility.default}
\alias{as.volatility.variance}
\title{volatility Class Coercing}
\usage{
as.volatility(x, ...)

\method{as.volatility}{default}(x, annualized = FALSE, dib = NULL, ...)

\method{as.volatility}{variance}(x, ...)
}
\arguments{
\item{x}{an object whose class will determine the method to be dispatched.
Value is the volatility or variance.}

\item{...}{further arguments to be passed to the next method, such as:}

\item{annualized}{for default method, TRUE if volatility value in x is annualized,
FALSE if volatility is daily.}

\item{dib}{days in one year, day count convention.}
}
\value{
All values of x as volatility class, according to given daily or annualized type.
}
\description{
Generic function for coercing to \code{"volatility"}. There is a default
method and a method for \code{"variance"}.
}
\examples{
as.volatility(1)
as.volatility(as.variance(.5))
as.volatility(0.2, TRUE, 252)
as.volatility(as.variance(10:30 / 100, TRUE))

}
\seealso{
\code{\link{as.variance}}.
}
