% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{bsmd1}
\alias{bsmd1}
\title{Computes d1 from Black-Scholes-Merton model.}
\usage{
bsmd1(spot, strike, time, rate, yield, sigma)
}
\arguments{
\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}

\item{sigma}{volatility of the stock price.}
}
\value{
Vector with d1.
}
\description{
d1 formula implementation, from Black-Scholes-Merton model.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\examples{
bsmd1(c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmd1(50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmd1(20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

}
