% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{bsmgreeks}
\alias{bsmgreeks}
\alias{bsmdelta}
\alias{bsmvega}
\alias{bsmgamma}
\alias{bsmtheta}
\alias{bsmrho}
\title{Computes the desired greek letter from Black-Scholes-Merton model.}
\usage{
bsmgreeks(
  greeks = c("delta", "gamma", "vega", "rho", "theta"),
  type,
  spot,
  strike,
  time,
  rate,
  yield,
  sigma
)

bsmdelta(type, spot, strike, time, rate, yield, sigma)

bsmvega(type, spot, strike, time, rate, yield, sigma)

bsmgamma(type, spot, strike, time, rate, yield, sigma)

bsmtheta(type, spot, strike, time, rate, yield, sigma)

bsmrho(type, spot, strike, time, rate, yield, sigma)
}
\arguments{
\item{greeks}{desired greek letter: delta, gamma, vega, rho or theta.}

\item{type}{call for call option, any other value for put.}

\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}

\item{sigma}{volatility of the stock price.}
}
\value{
Vector with desired greeks for input data.
}
\description{
Greeks formula implementation, from Black-Scholes-Merton model.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\section{bsmdelta}{

Delta greek letter measures the rate of change of the theoretical option
value with respect to changes in the underlying asset's price.
}

\section{bsmvega}{

Vega greek letter measures sensitivity of volatility.
}

\section{bsmgamma}{

Gamma greek letter measures the rate of change in the delta with respect
to changes in the underlying price.
}

\section{bsmtheta}{

Theta measures the sensitivity of the value of the derivative to the
passage of time.
}

\section{bsmrho}{

Rho measures sensitivity to the interest rate: it is the derivative
of the option value with respect to the risk free interest rate.
}

\examples{
bsmgreeks("delta", "put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

bsmdelta(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmdelta("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmdelta("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

bsmvega(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmvega("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmvega("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

bsmgamma(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmgamma("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmgamma("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

bsmtheta(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmtheta("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmtheta("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

bsmrho(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmrho("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmrho("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

}
