% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmimpvol.R
\name{bsmimpvol}
\alias{bsmimpvol}
\title{Computes implied volatility inverting Black-Scholes-Merton formula.}
\usage{
bsmimpvol(
  option_prices,
  type,
  spot,
  strike,
  time,
  rate,
  yield,
  tolerance = .Machine$double.eps,
  maxiter = 100,
  check = any
)
}
\arguments{
\item{option_prices}{stock option price.}

\item{type}{'call' for call option, any other value for put.}

\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}

\item{tolerance}{the tolerance for the approximation.}

\item{maxiter}{maximum number of iterations for bisection method.}

\item{check}{function for checking the convergence of all bisections.}
}
\value{
An approximation for the implied volatility. If the algorithm converges before
\code{maxiter} is reached, Black-Scholes-Merton formula calculated with this
volatility should not differ from \code{option_prices} by more than
\code{tolerance}.
}
\description{
Uses bisection method to calculate implied volatility, given other parameters
for Black-Scholes-Merton formula for pricing options on stocks.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\examples{
bsmimpvol(6, "call", 50, 52, 1, 0.1, 0)
bsmimpvol(c(3, 4, 5, 6), "call", 50, 52, 1, 0.1, 0)
bsmimpvol(1.5, "put", 40, 38, 0.5, 0.15, 0, tolerance = 1e-8)

}
