% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{bsmmoneyness}
\alias{bsmmoneyness}
\title{Computes a measure of moneyness for Black-Scholes-Merton model.}
\usage{
bsmmoneyness(spot, strike, time, rate, yield)
}
\arguments{
\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}
}
\value{
Vector with moneyness.
}
\description{
Moneyness formula implementation, for Black-Scholes-Merton model.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\examples{
bsmmoneyness(0:49, 1:50, 3, 0.1, 0)
bsmmoneyness(10, 10, 1, 0.12, 0:1000 / 1000)
bsmmoneyness(10, 10, 1, 0:1000 / 1000, 0)

}
