% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{bsmprice}
\alias{bsmprice}
\title{Computes option price via Black-Scholes-Merton formula.}
\usage{
bsmprice(type, spot, strike, time, rate, yield, sigma)
}
\arguments{
\item{type}{'call' for call option, any other value for put.}

\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}

\item{sigma}{volatility of the stock price.}
}
\value{
Vector with price of european options on stocks, according to
Black-Scholes-Merton model.
}
\description{
Black-Scholes-Merton formula implementation, pricing european options.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\examples{
bsmprice(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
bsmprice("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
bsmprice("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)

}
