% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{csm_obj_min_price}
\alias{csm_obj_min_price}
\alias{csm_obj_grad_price}
\title{Corrado-Su Model Objective Function in Price}
\usage{
csm_obj_min_price(
  par,
  type,
  spot,
  strike,
  rate,
  yield,
  time,
  price,
  weights = 1
)

csm_obj_grad_price(par, type, spot, strike, rate, time, price, weights)
}
\arguments{
\item{par}{a numeric value}

\item{type}{option type: \code{Call} or \code{Put}}

\item{spot}{a numeric value}

\item{strike}{a numeric value}

\item{rate}{a numeric value}

\item{yield}{a numeric value}

\item{time}{a numeric value}

\item{price}{a numeric value}

\item{weights}{a numeric value}
}
\description{
Objective function to Corrado-Su modelo that minimizes the error in price.
}
