% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{csm_obj_min_vol}
\alias{csm_obj_min_vol}
\title{Corrado-Su Model Objective Function in Implied Volatility}
\usage{
csm_obj_min_vol(par, type, spot, strike, rate, yield, time, vol, weights = 1)
}
\arguments{
\item{par}{a numeric value}

\item{type}{option type: \code{Call} or \code{Put}}

\item{spot}{a numeric value}

\item{strike}{a numeric value}

\item{rate}{a numeric value}

\item{yield}{a numeric value}

\item{time}{a numeric value}

\item{vol}{a numeric value}

\item{weights}{a numeric value}
}
\description{
Objective function to Corrado-Su modelo that minimizes the error in implied
volatility.
}
