% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{csmd}
\alias{csmd}
\title{Computes d from modified Corrado-Su model.}
\usage{
csmd(spot, strike, time, rate, yield, sigma, w)
}
\arguments{
\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}

\item{sigma}{volatility of the stock price.}

\item{w}{constant from modified Corrado-Su model.}
}
\value{
Vector with d.
}
\description{
d formula implementation from modified Corrado-Su model.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\examples{
csmd(14, 120:160 / 10, 0.5, 0.15, 0.1, 0.18, csmw(0.18, 0.5, 0, 3))
csmd(14, 13, 0.5, 0.15, 0.1, 0.18, csmw(0.18, 0.5, -1000:1000 / 100, 3))
csmd(14, 13, 0.5, 0.15, 0.1, 0.18, csmw(0.18, 0.5, 0, -1000:1000 / 100))

}
\seealso{
\code{\link{csmw}} for w from modified Corrado-Su model.\\
\code{\link{csmq3}} for Q_3 from modified Corrado-Su model.\\
\code{\link{csmq4}} for Q_4 from modified Corrado-Su model.\\
\code{\link{csmprice}} for modified Corrado-Su option pricing formula.
}
