% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{csmdelta}
\alias{csmdelta}
\title{Computes Delta from modified Corrado-Su model.}
\usage{
csmdelta(type, spot, strike, time, rate, yield, sigma, mu3, mu4)
}
\arguments{
\item{type}{'call' for call option, any other value for put.}

\item{spot}{current stock price.}

\item{strike}{the strike price.}

\item{time}{time to option expiration in years.}

\item{rate}{the risk-free interest rate.}

\item{yield}{the dividends that are expected to be paid.}

\item{sigma}{volatility of the stock price.}

\item{mu3}{skewness.}

\item{mu4}{kurtosis.}
}
\value{
Vector with options Delta.
}
\description{
This function computes delta greek for an option with the modified Corrado-Su
pricing model.
}
\examples{
csmdelta("call", 14, 13, 0.1, 0.15, 0, 0.5, 0.18, 3.53)
csmdelta("put", 14, 13, 0.1, 0.15, 0, 0.5, 0.18, 2.53)

}
