% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{csmw}
\alias{csmw}
\title{Computes w from modified Corrado-Su model.}
\usage{
csmw(sigma, time, mu3, mu4)
}
\arguments{
\item{sigma}{volatility of the stock price.}

\item{time}{time to option expiration in years.}

\item{mu3}{skewness.}

\item{mu4}{kurtosis.}
}
\value{
Vector with w.
}
\description{
w formula implementation from modified Corrado-Su model.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\examples{
csmw(0.25, 1.5, 0, -100:100 / 100)
csmw(0.25, 1.5, -100:100 / 100, 3)
csmw(0:100 / 100, 1.5, 0, 3)

}
\seealso{
\code{\link{csmd}} for d from modified Corrado-Su model.\\
\code{\link{csmq3}} for Q_3 from modified Corrado-Su model.\\
\code{\link{csmq4}} for Q_4 from modified Corrado-Su model.\\
\code{\link{csmprice}} for modified Corrado-Su option pricing formula.
}
