% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualize.R
\name{daily}
\alias{daily}
\alias{daily.volatility}
\alias{daily.variance}
\title{Daily Volatility or Variance}
\usage{
daily(x, ...)

\method{daily}{volatility}(x, ...)

\method{daily}{variance}(x, ...)
}
\arguments{
\item{x}{an object whose class will determine the method to be dispatched.
Volatility or variance, will become daily.}

\item{...}{further arguments to be passed to the next method.}
}
\value{
Same class with daily volatility or variance.
}
\description{
Transforms variance or volatility to daily. There is no default
method, so only classes \code{"volatility"}, \code{"variance"}, are supported.
}
\examples{
daily(annualize(as.variance(.5)))
daily(annualize(as.volatility(.25)))

}
\seealso{
\code{\link{annualize}} to transform variance or volatility to annual basis.
}
