% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiroot.R
\name{multiroot}
\alias{multiroot}
\title{Vectorial bisection method varying parameters from one function.}
\usage{
multiroot(
  func_,
  interval,
  ...,
  tolerance = .Machine$double.eps,
  maxiter = 100,
  check = any
)
}
\arguments{
\item{func_}{objective function for root finding.}

\item{interval}{a vector containing the end-points of the interval to be
searched for all the roots.}

\item{...}{additional named or unnamed arguments to be passed to func_,
may be vectorial, and this is the difference between multiroot and uniroot.}

\item{tolerance}{the desired accuracy (convergence tolerance).}

\item{maxiter}{the maximum number of iterations.}

\item{check}{function for checking convergence in \code{\link{any}} or
\code{\link{all}} of the roots. Since it is bisection method, they are
the same.}
}
\value{
Vector with roots.
}
\description{
Searches the interval for a root (i.e., zero) of the function func_ with
respect to its first argument. Other arguments may be vectorial, so
\code{multiroot} will return a vector with roots.
}
\details{
\code{multiroot} was created aiming implied volatility calculation for
several options.
}
\examples{
f <- function(x, y) x - y
multiroot(f, c(-3, 10), 1:3)
g <- function(sigma, premium, type, spot, strike, time, rate, yield) {
  bsmprice(type, spot, strike, time, rate, yield, sigma) - premium
}
multiroot(g, c(1e-8, 10), 5, "call", 30, 31, c(1, 1.1, 1.2, 1.3), 0.15, 0)
}
\seealso{
\code{\link{uniroot}} for R built-in one dimensional root finding.
}
