% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplib-package.R
\docType{package}
\name{oplib-package}
\alias{oplib}
\alias{oplib-package}
\title{oplib: Option Volatility and Pricing Models}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Many different option volatility models and option pricing models for Stocks, Interest Rates, and Currency European options in the Black-Scholes world.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilsonfreitas/oplib}
  \item Report bugs at \url{https://github.com/wilsonfreitas/oplib/issues}
}

}
\author{
\strong{Maintainer}: Wilson Freitas \email{wilson.freitas@gmail.com} (\href{https://orcid.org/0000-0002-0368-9267}{ORCID})

}
\keyword{internal}
