% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{print.volatility}
\alias{print.volatility}
\title{Special print method for volatility class.}
\usage{
\method{print}{volatility}(x, ...)
}
\arguments{
\item{x}{\code{"volatility"} class object to be printed.}

\item{...}{further arguments to be passed to the next method.}
}
\value{
invisible(x).
}
\description{
Makes objects of class \code{"volatility"} be printed differently.
}
\examples{
a <- as.volatility(1)
print(a)
a <- annualize(a)
print(a)

}
\seealso{
\code{\link{print.variance}}.
}
