% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi.R
\name{svi_var}
\alias{svi_var}
\title{SVI Variance}
\usage{
svi_var(a, b, m, rho, x, sigma)
}
\arguments{
\item{a}{'a' in the SVI model (~~ instantaneous variance)}

\item{b}{'b' in the SVI model (~~ speed of variance mean reversion)}

\item{m}{'m' in the SVI model (~~ long term expected variance)}

\item{rho}{'rho' in the SVI model (~~ correlation of variance and price)}

\item{x}{'x' in the SVI model, given by ln(K/F) where K is an options strike
price and F its underlying's future}

\item{sigma}{'sigma' in the SVI model (~~ volatility of variance)}
}
\value{
Volatility value as given by the SVI model
}
\description{
Calculates the variance as given by the SVI model for
the provided parameters.
}
\section{Recycle rule}{


These arguments handle the recycle rule so vectors can be provided
and once those vectors differs in length the recycle rule is applied.
}

\seealso{
\code{\link{svi_fit}} for SVI model fitting.
}
