% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{uncons-gram-charlier}
\alias{uncons-gram-charlier}
\alias{d_poly}
\alias{mu3}
\alias{mu4}
\alias{create_uncons_regionD}
\alias{uncons_regionD}
\title{Unconstrained Gram-Charlier Distribution Parametrization}
\usage{
d_poly(x)

mu3(x)

mu4(x)

create_uncons_regionD()

uncons_regionD(mu3p, mu4p)
}
\arguments{
\item{x}{a numeric value}

\item{mu3p}{a numeric value}

\item{mu4p}{a numeric value}
}
\description{
Creates the unconstrained parametrization for Gram-Charlier Distribution.
}
